moved solver
