+++ pages/02_Optimiser.py
@@ -7,7 +7,7 @@ import streamlit as st
 
 from ev_station_solver.constants import MOPTA_CONSTANTS
 from ev_station_solver.loading import load_locations
-from ev_station_solver.mopta_solver import MOPTASolver
+from ev_station_solver.solving.mopta_solver import MOPTASolver
 from ev_station_solver.streamlit import (
     CHARGER_BUILT_NAME,
     CHARGER_NOT_BUILT_NAME,
